create BookkeepingTable & refactor table code structure
